+++ src/components/Autocomplete/components/ComboBox/index.ts
@@ -1,2 +1,1 @@
 export * from './ComboBox';
-export * from './utilities';
