--- src/components/Autocomplete/components/ComboBox/utilities/combo-box/index.ts
@@ -1,3 +1,0 @@
-export type {ComboBoxTextFieldType, ComboBoxListBoxType} from './context';
-export {ComboBoxTextFieldContext, ComboBoxListBoxContext} from './context';
-export {useComboBoxTextField, useComboBoxListBox} from './hooks';
