--- src/components/Autocomplete/components/ComboBox/utilities/index.ts
@@ -1,1 +1,0 @@
-export * from './combo-box';
