+++ src/components/Autocomplete/components/ListBox/ListBox.tsx
@@ -13,11 +13,12 @@ import {useUniqueId} from '../../../../utilities/unique-id';
 import {Key} from '../../../../types';
 import {KeypressListener} from '../../../KeypressListener';
 import {VisuallyHidden} from '../../../VisuallyHidden';
-import {useComboBoxListBox} from '../ComboBox/utilities';
+import {useComboBoxListBox} from '../../../../utilities/combo-box';
+import {closestParentMatch} from '../../../../utilities/closest-parent-match';
+import {scrollIntoView} from '../../../../utilities/scroll-into-view';
+import {ListBoxContext} from '../../../../utilities/list-box';
+import type {NavigableOption} from '../../../../utilities/list-box';
 
-import {closestParentMatch} from './utilities/closest-parent-match';
-import {scrollIntoView} from './utilities/scroll-into-view';
-import {ListBoxContext} from './utilities/context/list-box';
 import {
   Option,
   Section,
@@ -27,7 +28,6 @@ import {
   TextOption,
   listBoxSectionDataSelector,
 } from './components';
-import type {NavigableOption} from './types';
 import styles from './ListBox.scss';
 
 export interface ListBoxProps {
