+++ src/components/Autocomplete/components/ListBox/components/Option/index.ts
@@ -1,2 +1,1 @@
-export {Option} from './Option';
-export type {OptionProps} from './Option';
+export * from './Option';
