+++ src/components/Autocomplete/components/ListBox/index.ts
@@ -1,4 +1,1 @@
-export {ListBox} from './ListBox';
-export type {ListBoxProps} from './ListBox';
-export {useListBox} from './utilities/hooks/useListBox';
-export * from './tests/utilities';
+export * from './ListBox';
