+++ src/components/Autocomplete/components/MappedOption/tests/MappedOption.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
+import {mountWithListBoxProvider} from 'test-utilities/list-box';
 
-import {ListBox, mountWithListBoxProvider} from '../../ListBox';
+import {ListBox} from '../../ListBox';
 import {MappedOption} from '../MappedOption';
 
 describe('MappedOption', () => {
