+++ src/components/Autocomplete/context.ts
@@ -1,0 +1,14 @@
+import {createContext} from 'react';
+
+interface MappedActionContextType {
+  role?: string;
+  url?: string;
+  external?: boolean;
+  onAction?(): void;
+  destructive?: boolean;
+  isAction: boolean;
+}
+
+export const MappedActionContext = createContext<MappedActionContextType>({
+  isAction: false,
+});
